+++ frontend/components/employee/DeliveryStatus.tsx
@@ -1,0 +1,50 @@
+import { Card, CardHeader, CardTitle, CardContent } from "@/components/ui/card";
+import { PieChart, Pie, Cell, Legend, ResponsiveContainer } from 'recharts';
+import { PieChartData } from "./types";
+
+interface DeliveryStatusProps {
+  data: PieChartData[];
+  totalOrders: number;
+}
+
+export const DeliveryStatus = ({ data, totalOrders }: DeliveryStatusProps) => {
+  return (
+    <Card className="bg-slate-900 border-slate-800">
+      <CardHeader>
+        <CardTitle className="text-xl text-white">Delivery Status</CardTitle>
+      </CardHeader>
+      <CardContent>
+        <div className="h-64">
+          <ResponsiveContainer width="100%" height="100%">
+            <PieChart>
+              <Pie
+                data={data}
+                cx="50%"
+                cy="50%"
+                innerRadius={60}
+                outerRadius={80}
+                paddingAngle={5}
+                dataKey="value"
+              >
+                {data.map((entry, index) => (
+                  <Cell key={`cell-${index}`} fill={entry.color} />
+                ))}
+              </Pie>
+              <Legend 
+                verticalAlign="bottom" 
+                height={36}
+                formatter={(value) => <span className="text-slate-300">{value}</span>}
+              />
+            </PieChart>
+          </ResponsiveContainer>
+        </div>
+
+        <div className="mt-4 space-y-2">
+          <div className="text-sm text-slate-400">
+            Total Orders: {totalOrders}
+          </div>
+        </div>
+      </CardContent>
+    </Card>
+  );
+};
