+++ frontend/components/employee/types.tsx
@@ -1,0 +1,22 @@
+export interface DeliveryOrder {
+    orderId: string;
+    orderName: string;
+    phoneNumber: string;
+    address: string;
+    isDelivered: boolean;
+    isCancelled?: boolean;
+    cancellationReason?: string;
+    items?: string[];
+  }
+  
+  export interface UndeliverableOrder {
+    orderId: string;
+    name: string;
+    phone: string;
+  }
+  
+  export interface PieChartData {
+    name: string;
+    value: number;
+    color: string;
+  }
